requirements.txt
python-telegram-bot==21.1
google-generativeai==0.8.3
supabase==2.4.3
python-dotenv==1.0.1
Pillow==10.3.0
httpx==0.27.0